+++ bgstally/activity.py
@@ -282,6 +282,9 @@ class Activity:
         self.dirty = True
         current_system = None
 
+        # Protect against rare case of null data, not able to trace how this can happen
+        if journal_entry.get('SystemAddress') == None or journal_entry.get('StarSystem') == None: return
+
         for system_address in self.systems:
             if system_address == str(journal_entry['SystemAddress']):
                 # We already have an entry for this system