+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## v4.0.0-xx - xxxx-xx-xx
 
+### Bug Fixes:
+
+* Fleet Carrier window was failing to initialise if the carrier's current system wasn't set.
 
 
 ## v4.0.0-a2 - 2024-05-16
@@ -18,7 +21,6 @@
 ### Bug Fixes:
 
 * 'Delete' button was hard to read when disabled on the CMDRs window.
-* Fleet Carrier window was failing to initialise if the carrier's current system wasn't set.
 
 
 ## v4.0.0-a1 - 2024-05-12