+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change Log
 
-## v4.0.0-xx - xxxx-xx-xx
+## v4.0.0-b1 - 2024-05-25
 
 ### Changes: